backend/requirements.txt
flask==3.0.0
flask-socketio==5.3.5
python-socketio==5.10.0
python-dotenv==1.0.0
pymongo==4.6.1
flask-cors==4.0.0
PyJWT==2.8.0
bcrypt==4.1.2